% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioural_obs.R
\name{make_combo}
\alias{make_combo}
\title{Make combo}
\usage{
make_combo(x, UL = "UL", LL = "LL", UR = "UR", LR = "LR", short)
}
\arguments{
\item{x}{a data.table}

\item{UL}{character vector, Upper Left}

\item{LL}{character vector, Lower Left}

\item{UR}{character vector, Upper Right}

\item{LR}{character vector,Lower Right}

\item{only_use}{character vector  UL, LL, UR or LR. if not missing will use this one only}
}
\description{
Make combo
}
\examples{
require(beR)
x = fread(
"UL; LL; UR; LR
M ; DG ; R,DG,R
M ; DG ; Y,R,DB
M ; DG ; DB,DG,DG
M ; DG ; Y,DB,R
M ; DG ; DB,R,O
M ; DG ; Y,DG,DG
M ; DG ; R,Y,O    
M ; DG ; NA    
"
, sep = ";" )
make_combo(x)
setnames(x, "LR", "LoR")
make_combo(x, LR = 'LoR')
make_combo(x, short = "UR")

}
