% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divorce.R
\name{pairing_status}
\alias{pairing_status}
\title{divorced pairs}
\usage{
pairing_status(d)
}
\description{
divorced pairs
}
\examples{
data(divorcees)
d = data.table(divorcees)
o = pairing_status(d)
merge(d, o, by = c('maleID', 'femaleID', 'season'))
}
