% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divorce.R
\name{pairing_status}
\alias{pairing_status}
\title{Divorced pairs}
\usage{
pairing_status(d)
}
\arguments{
\item{d}{data.frame containing maleID, femaleID and season}
}
\value{
a data.table similar to d but with two additional columns: male_status and female_status
}
\description{
Divorced pairs
}
\examples{
data(divorcees)
d = divorcees
x = pairing_status(divorcees)
x
}
